requirements.txt
dnspython>=2.2.1
requests==2.31.0